Moving classes to MVC packages
