+++ Jabberpoint/src/jabberPoint/model/BitmapItem.java
@@ -1,4 +1,4 @@
-package jabberPoint;
+package jabberPoint.model;
 import java.awt.Rectangle;
 import java.awt.Graphics;
 import java.awt.image.BufferedImage;
