+++ Jabberpoint/src/jabberPoint/model/DemoPresentation.java
@@ -1,4 +1,7 @@
-package jabberPoint;
+package jabberPoint.model;
+
+import jabberPoint.view.Accessor;
+
 /** Een ingebouwde demo-presentatie
  * @author Ian F. Darwin, ian@darwinsys.com, Gert Florijn, Sylvia Stuurman
  * @version 1.1 2002/12/17 Gert Florijn
