+++ Jabberpoint/src/jabberPoint/model/Presentation.java
@@ -1,5 +1,7 @@
-package jabberPoint;
+package jabberPoint.model;
 import java.util.ArrayList;
+
+import jabberPoint.view.SlideViewerComponent;
 
 
 /**
@@ -74,7 +76,7 @@ public class Presentation {
 	}
 
 	// Verwijder de presentatie, om klaar te zijn voor de volgende
-	void clear() {
+	public void clear() {
 		showList = new ArrayList<Slide>();
 		setSlideNumber(-1);
 	}
