+++ Jabberpoint/src/jabberPoint/model/SlideItem.java
@@ -1,4 +1,4 @@
-package jabberPoint;
+package jabberPoint.model;
 import java.awt.Rectangle;
 import java.awt.Graphics;
 import java.awt.image.ImageObserver;
