+++ Jabberpoint/src/jabberPoint/model/TextItem.java
@@ -1,4 +1,4 @@
-package jabberPoint;
+package jabberPoint.model;
 import java.awt.Rectangle;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
