+++ Jabberpoint/src/jabberPoint/view/AboutBox.java
@@ -1,4 +1,4 @@
-package jabberPoint;
+package jabberPoint.view;
 import java.awt.Frame;
 import javax.swing.JOptionPane;
